Update and rename ip-rotation.py to SpammingSpider.py
The above script is a Python code that rotates IP addresses and checks the reputation of each IP before sending a GET request to a website (http://example.com) [Change this example to the actual website]. The IP addresses to be used are listed in the ip_list array. For each IP, the script checks if it is already in the proxy list, and if not, it sets the proxy and checks the reputation of the IP using the check_ip_reputation function (which is currently not implemented). If the reputation is not good, the IP is skipped. If the reputation is good, the script sends a GET request to the website, retrying up to 3 times if necessary in case of timeouts or other errors. The status code of the response is logged.