Update and rename password-dictionary.py to PasswordPirate.py
The above code is a password generator that creates a list of possible password options based on input words and special characters. The user is prompted to enter a list of words separated by commas, as well as the maximum length of the password. The code checks to make sure the input words contain only alphanumeric characters and the max length is a positive integer. The password options are generated by adding special characters to the beginning, end, and middle of the words, repeating the words, and combining the words. The list of passwords is then made unique by removing duplicates. The generated passwords are printed for the user to use.